% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroSurvSizePred-package.R
\docType{package}
\name{HydroSurvSizePred-package}
\alias{HydroSurvSizePred}
\alias{HydroSurvSizePred-package}
\title{HydroSurvSizePred: Salmon smolt length, predation risk, and survival}
\description{
App used to explore sp/su Chinook Salmon smolt length distributions, predator risk from Northern Pikeminnow and Pacific Hake, and survival across years and seasons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/caitobrien/HydroSurvSizePred.git}
  \item Report bugs at \url{https://github.com/caitobrien/HydroSurvSizePred.git/issues}
}

}
\author{
\strong{Maintainer}: CAITLIN OBRIEN \email{CSOBRIEN@UW.EDU}

}
\keyword{internal}
